requirements.txt
# API & server
fastapi
uvicorn[standard]

# LangChain
langchain
langchain-core
langchain-community
langchain-chroma
langchain-groq
langchain-text-splitters

# Embeddings (local, no API key needed)
sentence-transformers

# Vector store
chromadb

# Document loaders
pypdf
beautifulsoup4
requests

# Env
python-dotenv

# Frontend
streamlit
